+++ app/Http/Controllers/MuzakiController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use DB;
 use App\Muzaki;
+use App\Pengguna;
 use Illuminate\Http\Request;
 
 class MuzakiController extends Controller
@@ -13,6 +15,11 @@ class MuzakiController extends Controller
         $data['status'] = 200;
         $data['data'] = Muzaki::all();
 
+        $data = DB::select("SELECT * FROM muzakis LEFT JOIN kantors ON muzakis.id_kantor = kantors.id_kantor LEFT JOIN penggunas ON muzakis.id_pengguna = Penggunas.id_pengguna");
+        // $data = DB::table("muzakis")
+        //     ->leftjoin("kantors", "muzakis.id_kantor", "=", "kantors.id_kantor")
+        //     ->select("muzakis.alamat_muzaki as alamat_muzaki",  "kantors.alamat_muzaki as alamat_kantor")
+        //     ->get();
         return $data;
     }
 
@@ -24,13 +31,13 @@ class MuzakiController extends Controller
         $muzaki->nik = $request->nik;
         $muzaki->nama_muzaki = $request->nama_muzaki;
         $muzaki->jk = $request->jk;
-        $muzaki->alamat = $request->alamat;
+        $muzaki->alamat_muzaki = $request->alamat_muzaki;
         $muzaki->profesi = $request->profesi;
-        $muzaki->no_hp = $request->no_hp;
-        $muzaki->kategori = $request->kategori;
-        $muzaki->status = $request->status;
-        $muzaki->kode_petugas = $request->kode_petugas;
-        $muzaki->no_kantor = $request->no_kantor;
+        $muzaki->telepon_muzaki = $request->telepon_muzaki;
+        $muzaki->kategori_muzaki = $request->kategori_muzaki;
+        $muzaki->status_muzaki = 1;
+        $muzaki->id_pengguna = $request->id_pengguna;
+        $muzaki->id_kantor = $request->id_kantor;
 
         $simpan = $muzaki->save();
         if ($simpan) {
@@ -52,35 +59,43 @@ class MuzakiController extends Controller
         $nik = $request->nik;
         $nama_muzaki = $request->nama_muzaki;
         $jk = $request->jk;
-        $alamat = $request->alamat;
+        $alamat_muzaki = $request->alamat_muzaki;
         $profesi = $request->profesi;
-        $no_hp = $request->no_hp;
-        $kategori = $request->kategori;
-        $status = $request->status;
-        $kode_petugas = $request->kode_petugas;
-        $no_kantor = $request->no_kantor;
+        $telepon_muzaki = $request->telepon_muzaki;
+        $kategori_muzaki = $request->kategori_muzaki;
+        $status_muzaki = $request->status_muzaki;
+        $id_pengguna = $request->id_pengguna;
+        $id_kantor = $request->id_kantor;
 
 
         $muzaki = Muzaki::find($id);
-        $muzaki->npwz = $npwz;
-        $muzaki->nik = $nik;
-        $muzaki->nama_muzaki = $nama_muzaki;
-        $muzaki->jk = $jk;
-        $muzaki->alamat = $alamat;
-        $muzaki->profesi = $profesi;
-        $muzaki->no_hp = $no_hp;
-        $muzaki->kategori = $kategori;
-        $muzaki->kode_petugas = $kode_petugas;
-        $muzaki->no_kantor = $no_kantor;
 
-        $update = $muzaki->update();
-        if ($update) {
-            $data['status'] = true;
-            $data['message'] = "Berhasil di Update ";
+        if ($muzaki) {
+            # code...
+            $muzaki->npwz = $npwz;
+            $muzaki->nik = $nik;
+            $muzaki->nama_muzaki = $nama_muzaki;
+            $muzaki->jk = $jk;
+            $muzaki->alamat_muzaki = $alamat_muzaki;
+            $muzaki->profesi = $profesi;
+            $muzaki->telepon_muzaki = $telepon_muzaki;
+            $muzaki->kategori_muzaki = $kategori_muzaki;
+            $muzaki->status_muzaki = $status_muzaki;
+            $muzaki->id_pengguna = $id_pengguna;
+            $muzaki->id_kantor = $id_kantor;
+
             $data['data'] = $muzaki;
+            $update = $muzaki->update();
+            if ($update) {
+                $data['status'] = true;
+                $data['message'] = "Berhasil di Update ";
+                $data['data'] = $muzaki;
+            } else {
+                $data['status'] = false;
+                $data['message'] = "Gagal di Update ";
+                $data['data'] = null;
+            }
         } else {
-            $data['status'] = false;
-            $data['message'] = "Gagal di Update ";
             $data['data'] = null;
         }
         return $data;
@@ -90,15 +105,21 @@ class MuzakiController extends Controller
     public function delete($id)
     {
         $muzaki = Muzaki::find($id);
-        $delete = $muzaki->delete();
 
-        if ($delete) {
-            $data['status'] = true;
-            $data['message'] = "Data Berhasil di Hapus ";
-            $data['data'] = $muzaki;
+        if ($muzaki) {
+            # code...
+            $delete = $muzaki->delete();
+
+            if ($delete) {
+                $data['status'] = true;
+                $data['message'] = "Data Berhasil di Hapus ";
+                $data['data'] = $muzaki;
+            } else {
+                $data['status'] = false;
+                $data['message'] = "Data Gagal di Hapus ";
+                $data['data'] = null;
+            }
         } else {
-            $data['status'] = false;
-            $data['message'] = "Data Gagal di Hapus ";
             $data['data'] = null;
         }
